pyproject.toml
[project]
name = "lang-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.24.0",
    "langchain[google-genai]>=0.3.25",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
]
